Relocating the Sync related files to components
