+++ renderer/pages/sync.js
@@ -1,4 +1,4 @@
-import Sync from '@/modules/projects/Sync';
+import Sync from '@/components/Sync/Sync';
 import AuthenticationContextProvider from '@/components/Login/AuthenticationContextProvider';
 import AutographaContextProvider from '@/components/context/AutographaContext';
 import SyncContextProvider from '@/components/Sync/SyncContextProvider';
