+++ renderer/src/components/Sync/Gitea/ProjectMerge/ProjectMergePopUp.js
@@ -3,9 +3,9 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import ConfirmationModal from '@/layouts/editor/ConfirmationModal';
 import PropTypes from 'prop-types';
-import PopUpModal from '@/layouts/Sync/PopUpModal';
 import { SnackBar } from '@/components/SnackBar';
 import useAddNotification from '@/components/hooks/useAddNotification';
+import PopUpModal from '../../PopUpModal';
 import { VerticalLinearStepper } from '../../VerticalStepperProgress';
 import { mergeProgressSteps } from './ProjectMergeConst';
 import LoadingSpinner from '../../LoadingSpinner';
